Ported one more test
